:hammer: Fix: 여러가지 버그 수정

+++ backend/src/graphql/resolvers/AnswerResolver.ts
@@ -1,5 +1,6 @@
 import {
   Arg,
+  Ctx,
   FieldResolver,
   Int,
   Mutation,
@@ -13,21 +14,27 @@ import AnswerInput from "../inputTypes/AnswerInput";
 import PostService from "../services/PostService";
 import UserService from "../services/UserService";
 
+const getUserId = (headers: any): number => {
+  if (!headers.authorization) throw new Error("Auth Error");
+  const token = headers.authorization.split(" ")[1];
+  return (verify(token, "jwtprivate") as any).userId;
+};
+
 @Resolver(PostAnswer)
 export default class AnswerResolver {
   @Mutation(() => PostAnswer, { description: "답변글 작성 Mutation" })
   async addNewAnswer(
     @Arg("questionId", () => Int, { description: "질문글 ID" })
     questionId: number,
-    @Arg("accessToken", () => String, { description: "access token" })
-    accessToken: string,
-    @Arg("data") answerData: AnswerInput
+    @Arg("data") answerData: AnswerInput,
+    @Ctx("headers") headers: any
   ): Promise<PostAnswer> {
-    const id = verify(accessToken, "jwtprivate") as string;
+    const userId = getUserId(headers);
+    console.log(userId);
     const newAnswer = await PostService.addNewAnswer(
       answerData,
       {
-        id: Number(id),
+        id: userId,
       },
       questionId
     );
@@ -46,8 +53,13 @@ export default class AnswerResolver {
     @Arg("answernId", () => Int, { description: "수정할 답변글의 ID" })
     answerId: number,
     @Arg("data", { description: "수정할 답변글 내용" })
-    answerInput: AnswerInput
+    answerInput: AnswerInput,
+    @Ctx("headers") headers: any
   ): Promise<PostAnswer> {
+    const userId = getUserId(headers);
+    const answer = await PostService.findOneAnswerById(answerId);
+    const anwerAuthorId = answer.userId;
+    if (userId !== anwerAuthorId) throw new Error("Not your Post!");
     const updateResult = await PostService.updateAnswer(answerId, answerInput);
 
     return await PostService.findOneAnswerById(answerId);
@@ -57,8 +69,13 @@ export default class AnswerResolver {
     description: "답변글 삭제 Mutation, 삭제 여부를 Boolean 으로 반환합니다.",
   })
   async deleteAnswer(
-    @Arg("answerId", { description: "삭제할 질문글의 ID" }) answerId: number
+    @Arg("answerId", { description: "삭제할 질문글의 ID" }) answerId: number,
+    @Ctx("headers") headers: any
   ): Promise<boolean> {
+    const userId = getUserId(headers);
+    const answer = await PostService.findOneAnswerById(answerId);
+    const anwerAuthorId = answer.userId;
+    if (userId !== anwerAuthorId) throw new Error("Not your Post!");
     const isDeleted = await PostService.deleteAnswer(answerId);
 
     return isDeleted;
